requirements.txt
Django==4.0.6
django-mysql==4.7.0
mysqlclient==2.1.1
celery==5.2.7
redis==4.3.4
flower==1.1.0
pytest==7.1.2
pytest-django==4.5.2
django-environ==0.9.0
google-api-python-client==2.55.0
djangorestframework==3.13.1